renaming edge to edge_cli for command line version
